+++ src/lib/Prefs/PrefsKeys.js
@@ -291,6 +291,18 @@ var PrefsKeys = class
 
         this.setKey(
             'visibility',
+            'calendar',
+            'GtkSwitch',
+            true,
+            {
+                default: true,
+                minimal: true,
+                superminimal: true,
+            }
+        );
+
+        this.setKey(
+            'visibility',
             'events-button',
             'GtkSwitch',
             true,