+++ CHANGELOG.md
@@ -18,6 +18,7 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 ### Added
 
 - Events visibility in clock menu.
+- Calendar visibility in clock menu.
 - Dutch translation by [@Vistaus](https://gitlab.gnome.org/Vistaus).
 
 ## [19.0.0 Ancher] - 2022-03-02